refactor the project
