+++ src/Pages/Home.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import InputForm from '../components/Books/InputForm';
+import AllBookListedBooks from '../components/Books/AllBooks';
+
+const AllBooks = () => (
+  <section>
+    <div>
+      <AllBookListedBooks />
+      <InputForm />
+    </div>
+  </section>
+);
+
+export default AllBooks;
