+++ src/components/Books/AllBooks.jsx
@@ -1,23 +1,21 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import InputForm from '../components/Books/InputForm';
-import ShowBook from '../components/Books/ShowBook';
+import SingleBook from './SingleBook';
 
-const AllBooks = () => {
+const ShowBook = () => {
   const books = useSelector((state) => state.books.books);
   return (
     <div>
       {books.map((book) => (
-        <ShowBook
+        <SingleBook
           key={book.id}
           title={book.title}
           author={book.author}
           category={book.category}
         />
       ))}
-      <InputForm />
     </div>
   );
 };
 
-export default AllBooks;
+export default ShowBook;
